Intro endpoints

+++ src/main/java/org/example/api/data/entity/Card.java
@@ -9,7 +9,7 @@ import java.util.Date;
 public class Card {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long cardId;
+    private Integer cardId;
 
     private String type;
     private Long number;
